# api-document.yaml
openapi: "3.0.2"
info:
  title: "Sample API"
  version: "1.0.0"
servers:
  - url: http://localhost:8080/
paths:
...
  /posts:
      post:
        summary: "Create post"
        operationId: "createPost"
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  content:
                    type: string
                  category:
                    type: string
        responses:
          "200":
            description: "Post created"
          "400":
            description: "Invalid input"
      get:
        summary: "Get list of posts"
        operationId: "getPosts"
        responses:
          "200":
            description: "A list of posts"
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    posts:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          userId:
                            type: integer
                          content:
                            type: string
                          category:
                            type: string
                          createdAt:
                            type: string
                            format: date-time
                          updatedAt:
                            type: string
                            format: date-time
...